Allow an assignment to have fractional points

+++ src/_reactComponents/Activity/SettingComponents.jsx
@@ -571,8 +571,8 @@ export const TotalPointsOrPercent = ({ courseId, doenetId }) => {
                 setTotalPointsOrPercent(0);
                 totalPointsOrPercentLocal = 0;
               } else {
-                totalPointsOrPercentLocal = parseInt(totalPointsOrPercent);
-                setTotalPointsOrPercent(parseInt(totalPointsOrPercent));
+                totalPointsOrPercentLocal = parseFloat(totalPointsOrPercent);
+                setTotalPointsOrPercent(parseFloat(totalPointsOrPercent));
               }
 
               updateAssignmentSettings(doenetId, {
